HitObjectType -> LegacyHitObjectType, move to Legacy namespace.
