+++ osu.Game/Modes/Objects/LegacyHitObjectParser.cs
@@ -16,13 +16,13 @@ namespace osu.Game.Modes.Objects
         public override HitObject Parse(string text)
         {
             string[] split = text.Split(',');
-            var type = (HitObjectType)int.Parse(split[3]) & ~HitObjectType.ColourHax;
-            bool combo = type.HasFlag(HitObjectType.NewCombo);
-            type &= ~HitObjectType.NewCombo;
+            var type = (LegacyHitObjectType)int.Parse(split[3]) & ~LegacyHitObjectType.ColourHax;
+            bool combo = type.HasFlag(LegacyHitObjectType.NewCombo);
+            type &= ~LegacyHitObjectType.NewCombo;
 
             HitObject result;
 
-            if ((type & HitObjectType.Circle) > 0)
+            if ((type & LegacyHitObjectType.Circle) > 0)
             {
                 result = new LegacyHit
                 {
@@ -30,7 +30,7 @@ namespace osu.Game.Modes.Objects
                     NewCombo = combo
                 };
             }
-            else if ((type & HitObjectType.Slider) > 0)
+            else if ((type & LegacyHitObjectType.Slider) > 0)
             {
                 CurveType curveType = CurveType.Catmull;
                 double length = 0;
@@ -85,14 +85,14 @@ namespace osu.Game.Modes.Objects
                     NewCombo = combo
                 };
             }
-            else if ((type & HitObjectType.Spinner) > 0)
+            else if ((type & LegacyHitObjectType.Spinner) > 0)
             {
                 result = new LegacySpinner
                 {
                     EndTime = Convert.ToDouble(split[5], CultureInfo.InvariantCulture)
                 };
             }
-            else if ((type & HitObjectType.Hold) > 0)
+            else if ((type & LegacyHitObjectType.Hold) > 0)
             {
                 // Note: Hold is generated by BMS converts
                 result = new LegacyHold
