+++ osu.Game/osu.Game.csproj
@@ -113,7 +113,7 @@
     <Compile Include="Modes\Objects\Types\IHasRepeats.cs" />
     <Compile Include="Modes\Objects\Types\IHasPosition.cs" />
     <Compile Include="Modes\Objects\Types\IHasSpecial.cs" />
-    <Compile Include="Modes\Objects\Types\LegacyHitObjectType.cs" />
+    <Compile Include="Modes\Objects\Legacy\LegacyHitObjectType.cs" />
     <Compile Include="Modes\Replay.cs" />
     <Compile Include="Modes\Score.cs" />
     <Compile Include="Modes\ScoreProcesssor.cs" />
